formulário para estudantes atualizarem projetos

+++ projetos/sites.py
@@ -17,8 +17,7 @@ from django.shortcuts import get_object_or_404
 from django.core.files.uploadedfile import InMemoryUploadedFile, TemporaryUploadedFile
 from django.http import HttpResponse
 from django.contrib.auth.decorators import login_required
-from django.shortcuts import render
-
+from django.shortcuts import render, redirect
 from users.models import Alocacao
 from projetos.models import Projeto, Configuracao
 
@@ -54,29 +53,52 @@ def sites(request, projeto_id, path):
 
 
 @login_required
-def upload_site(request, projeto_id):
+def upload_site_e_pastas(request, projeto_id):
+    
     if request.method == "POST":
-        zip_file = request.FILES["zipsite"]
-        if isinstance(zip_file, (InMemoryUploadedFile, TemporaryUploadedFile)):
-            configuracao = get_object_or_404(Configuracao)
-            MAX_SIZE = configuracao.maxMB_filesize * 1048576
-            with zipfile.ZipFile(zip_file, 'r') as zip_ref:
-                total_size = sum((file.file_size for file in zip_ref.infolist()))
-                if total_size > MAX_SIZE:
-                    return HttpResponse("Arquivo descomprimido maior que o limite: " + str(configuracao.maxMB_filesize) + "MB.")
-                projeto = get_object_or_404(Projeto, id=projeto_id)
-                site_root = settings.SITE_ROOT + "/projeto"+str(projeto.id)
-                if os.path.exists(site_root):
-                    os.system("rm -rf " + site_root)
-                zip_ref.extractall(site_root)  # extract the zip file to the project path
-                mensagem = "Site atualizado com sucesso."
-                mensagem += "<br><a href='/sites/"+str(projeto_id)+"/'>Visualizar site</a>"
-                context = {
-                    "area_principal": True,
-                    "mensagem": mensagem,
-                }
-                return render(request, "generic.html", context=context)
-        else:
-            return HttpResponse("Não é arquivo.")
+        projeto = get_object_or_404(Projeto, id=projeto_id)
+
+        if "Carregar" in request.POST:
+            if "zipsite" not in request.FILES:
+                return HttpResponse("Nenhum arquivo foi enviado.")
+            zip_file = request.FILES["zipsite"]
+            if isinstance(zip_file, (InMemoryUploadedFile, TemporaryUploadedFile)):
+                configuracao = get_object_or_404(Configuracao)
+                MAX_SIZE = configuracao.maxMB_filesize * 1048576
+                with zipfile.ZipFile(zip_file, 'r') as zip_ref:
+                    total_size = sum((file.file_size for file in zip_ref.infolist()))
+                    if total_size > MAX_SIZE:
+                        return HttpResponse("Arquivo descomprimido maior que o limite: " + str(configuracao.maxMB_filesize) + "MB.")
+                    site_root = settings.SITE_ROOT + "/projeto"+str(projeto.id)
+                    if os.path.exists(site_root):
+                        os.system("rm -rf " + site_root)
+                    zip_ref.extractall(site_root)  # extract the zip file to the project path
+                    projeto.site = "/sites/"+str(projeto_id)+"/"
+                    projeto.save()
+            else:
+                return HttpResponse("Não é arquivo.")
         
-    return HttpResponse("Invalid request.")
+        if "Remover" in request.POST:
+            site_root = settings.SITE_ROOT + "/projeto"+str(projeto.id)
+            if os.path.exists(site_root):
+                os.system("rm -rf " + site_root)
+            projeto.site = None
+            projeto.save()
+
+        if "Salvar" in request.POST:
+            if "pastas_do_projeto" in request.POST:
+                pastas_do_projeto = request.POST.get("pastas_do_projeto", None)
+                projeto.pastas_do_projeto = pastas_do_projeto
+                projeto.save()
+            if "site" in request.POST:
+                site = request.POST.get("site", "").strip()
+                if site:
+                    if site[:4] == "http":
+                        projeto.site = site
+                    else:
+                        projeto.site = "http://" + site
+                else:
+                    projeto.site = None
+            projeto.save()
+        
+    return redirect("/projetos/meuprojeto")
